% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_transformation.R
\name{clr_trans}
\alias{clr_trans}
\title{Centered log ratio transform}
\usage{
clr_trans(data)
}
\arguments{
\item{data}{The data matrix (rows are samples and columns are features/variables)}
}
\value{
The transformed data matrix
}
\description{
Compute the centered log ratio transform of a data matrix
}
