% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{completed_counter}
\alias{completed_counter}
\title{Progress bar for long running tasks}
\usage{
completed_counter(fn, total)
}
\arguments{
\item{fn}{function to wrap}

\item{total}{total number of iterations}
}
\value{
wrapped function
}
\description{
Progress bar for long running tasks
}
