% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lib_size.R
\name{create_lib_size}
\alias{create_lib_size}
\title{Create library size for a dataset at one timepoint or multiple timepoints}
\usage{
create_lib_size(data)
}
\arguments{
\item{data}{A 3D or 2D array of counts or transformed data with dimensions samples x taxa}
}
\value{
Library size across samples and optionally timepoints
}
\description{
Create library size for a dataset at one timepoint or multiple timepoints
}
