% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_dis.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance matrix}
\usage{
distance_matrix(network_ls)
}
\arguments{
\item{network_ls}{network list}
}
\value{
data frame with IVI values
}
\description{
Distance matrix
}
