% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdimApprox_single.R
\name{ICA_approx}
\alias{ICA_approx}
\title{ICA approximation}
\usage{
ICA_approx(X, ncomp = 1)
}
\arguments{
\item{X}{The data matrix}

\item{ncomp}{The number of components to use for ICA}
}
\value{
The low dimensional data matrix using ICA
}
\description{
ICA approximation
}
