% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influential_nodes.R
\name{IVI_values}
\alias{IVI_values}
\title{IVI values}
\usage{
IVI_values(network_ls, group, days)
}
\arguments{
\item{network_ls}{network list}

\item{group}{group name}

\item{days}{days}
}
\value{
data frame with IVI values
}
\description{
IVI values
}
