% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE_boot.R
\name{LUPINE_bootsrap}
\alias{LUPINE_bootsrap}
\title{LUPINE_bootsrap function}
\usage{
LUPINE_bootsrap(
  data,
  day_range = NULL,
  is.transformed = FALSE,
  lib_size = NULL,
  ncomp = 1,
  single = FALSE,
  singleMethod = "pca",
  excluded_taxa = NULL,
  cutoff = 0.05,
  nboot = 1000
)
}
\arguments{
\item{data}{The data matrix (rows are samples, columns are features/variables, and slices are time points)}

\item{day_range}{day_range The range of days to perform LUPINE bootstrap on}

\item{is.transformed}{A logical indicating whether the data is transformed or not}

\item{lib_size}{A matrix of library sizes for each sample and time point}

\item{ncomp}{The number of components to use for dimensionality reduction}

\item{single}{A logical indicating whether to use LUPINE for a single time point or longitudinal data}

\item{singleMethod}{The method to use for dimensionality reduction for single time point}

\item{excluded_taxa}{A list of taxa to be excluded at each time point}

\item{cutoff}{cutoff The cutoff value for the p-value to determine significance of the correlation}

\item{nboot}{The number of bootstrap iterations to perform}
}
\value{
A list of correlations and p-values for all days
}
\description{
LUPINE_bootsrap function
}
