% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE_longitudinal.R
\name{LUPINE_longitudinal}
\alias{LUPINE_longitudinal}
\title{LUPINE for longitudinal data}
\usage{
LUPINE_longitudinal(
  data,
  day,
  excluded_taxa = NULL,
  is.transformed = FALSE,
  lib_size = NULL,
  ncomp = 1
)
}
\arguments{
\item{data}{A 3D array of counts or transformed data with dimensions samples x taxa x time points}

\item{day}{The time point for which the correlations are calculated}

\item{excluded_taxa}{A list of taxa to be excluded at each time point}

\item{is.transformed}{A logical indicating whether the data is transformed or not}

\item{lib_size}{A matrix of library sizes for each sample and time point}

\item{ncomp}{The number of components to use for dimensionality reduction}
}
\value{
A list of correlations and p-values
}
\description{
LUPINE for longitudinal data
}
