% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE_single.R
\name{LUPINE_single}
\alias{LUPINE_single}
\title{LUPINE for single time point}
\usage{
LUPINE_single(
  data,
  day,
  excluded_taxa = NULL,
  is.transformed = FALSE,
  lib_size = NULL,
  method = "pca",
  ncomp = 1
)
}
\arguments{
\item{data}{A 3D array of counts or transformed data with dimensions samples x taxa x time points}

\item{day}{The time point for which the correlations are calculated}

\item{excluded_taxa}{A list of taxa to be excluded at each time point}

\item{is.transformed}{A logical indicating whether the data is transformed or not}

\item{lib_size}{A matrix of library sizes for each sample and time point}

\item{method}{The method to use for dimensionality reduction. Options are "pca", "ica", "rpca"}

\item{ncomp}{The number of components to use for dimensionality reduction}
}
\value{
A list of correlations and p-values
}
\description{
LUPINE for single time point
}
