% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE_t.R
\name{LUPINE_t}
\alias{LUPINE_t}
\title{LUPINE for longitudinal data}
\usage{
LUPINE_t(
  data,
  day_index,
  num_lags = 999,
  excluded_taxa = NULL,
  is.transformed = FALSE,
  lib_size = NULL,
  ncomp = 1
)
}
\arguments{
\item{data}{A 3D array of counts or transformed data with dimensions samples x taxa x time points}

\item{day_index}{The index of the time point for which the correlations are calculated}

\item{num_lags}{The number of previous time points to be used for the calculation of the correlation}

\item{excluded_taxa}{A list of taxa to be excluded at each time point}

\item{is.transformed}{A logical indicating whether the data is transformed or not}

\item{lib_size}{A matrix of library sizes for each sample and time point}

\item{ncomp}{The number of components to use for dimensionality reduction}
}
\value{
A list of correlations and p-values
}
\description{
LUPINE for longitudinal data
}
