% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkTest.R
\name{MantelTest_matrix}
\alias{MantelTest_matrix}
\title{Network Mantel test}
\usage{
MantelTest_matrix(network_ls)
}
\arguments{
\item{network_ls}{network list}
}
\value{
matrix of p-values
}
\description{
Network Mantel test
}
