% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdimApprox_single.R
\name{PCA_approx}
\alias{PCA_approx}
\title{PCA approximation}
\usage{
PCA_approx(X, ncomp = 1)
}
\arguments{
\item{X}{The data matrix}

\item{ncomp}{The number of components to use for PCA}
}
\value{
The low dimensional data matrix using PCA
}
\description{
PCA approximation
}
