% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdimApprox_longitudinal.R
\name{PLS_approx}
\alias{PLS_approx}
\title{PLS low dimensional approximaton for longitudinal data}
\usage{
PLS_approx(X_array, day, ncomp = 1, taxa_names = NULL)
}
\arguments{
\item{X_array}{The data array}

\item{day}{The day to approximate}

\item{ncomp}{The number of components to use for PLS}

\item{taxa_names}{The taxa names to use for PLS}
}
\value{
The low dimensional data matrix using PLS
}
\description{
PLS low dimensional approximaton for longitudinal data
}
