% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowdimApprox_single.R
\name{RPCA_approx}
\alias{RPCA_approx}
\title{RPCA approximation}
\usage{
RPCA_approx(X, ncomp = 1)
}
\arguments{
\item{X}{The data matrix}

\item{ncomp}{The number of components to use for RPCA}
}
\value{
The low dimensional data matrix Uusing RPCA
}
\description{
RPCA approximation
}
