% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE.R
\name{LUPINE}
\alias{LUPINE}
\title{LUPINE function}
\usage{
LUPINE(
  data,
  is.transformed = FALSE,
  lib_size = NULL,
  ncomp = 1,
  single = FALSE,
  singleMethod = "pca",
  excluded_taxa = NULL,
  cutoff = 0.05
)
}
\arguments{
\item{data}{The data matrix (rows are samples, columns are features/variables, and slices are time points)}

\item{is.transformed}{A logical indicating whether the data is transformed or not}

\item{lib_size}{A matrix of library sizes for each sample and time point}

\item{ncomp}{The number of components to use for dimensionality reduction}

\item{single}{A logical indicating whether to use LUPINE for a single time point or longitudinal data}

\item{singleMethod}{The method to use for dimensionality reduction for single time point}

\item{excluded_taxa}{A list of taxa to be excluded at each time point}

\item{cutoff}{The cutoff value for the p-value to determine significance of the correlation}
}
\value{
A list of correlations and p-values for all days
}
\description{
LUPINE function
}
