% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUPINE_single_timepoint.R
\name{LUPINE_single_timepoint}
\alias{LUPINE_single_timepoint}
\title{LUPINE for single time point}
\usage{
LUPINE_single_timepoint(
  data_timepoint,
  lib_size = NULL,
  method = "pca",
  ncomp = 1
)
}
\arguments{
\item{lib_size}{A matrix of library sizes for each sample and time point, optional. Currently if used regression models will be run on log+1 and libsize accounted for}

\item{method}{The method to use for dimensionality reduction. Options are "pca", "ica", "rpca"}

\item{ncomp}{The number of components to use for dimensionality reduction}

\item{data}{A 2D array of counts or transformed data with dimensions samples x taxa}
}
\value{
A list of correlations and p-values
}
\description{
LUPINE for single time point
}
